refactor: remove inline styles from edit page and extract shared export util - move toggle switch, delete section divider, and confirm row inline styles to CSS classes - extract duplicated handleExport logic into an exportUtils file
